+++ src/index.js
@@ -1,3 +1,1 @@
-import Container from './container';
-
-export default Container;
+export default from './container';